Corrected target to dtype float

+++ pegasus/generate_data.py
@@ -151,9 +151,10 @@ def generate_data(tokenizer: PreTrainedTokenizer,
     test_data_numpy = np.array(test_data)
     
     # indices for one_hot must be dtype torch.int64
-    training_targets = one_hot(torch.as_tensor(np.argmax(training_data_numpy, axis=1), dtype=torch.int64), datapoint_length).to(device)
+    # targets with class probabilities must be a floating type
+    training_targets = one_hot(torch.as_tensor(np.argmax(training_data_numpy, axis=1), dtype=torch.int64), datapoint_length).to(torch.float32).to(device)
     
-    test_targets = one_hot(torch.as_tensor(np.argmax(test_data_numpy, axis=1), dtype=torch.int64), datapoint_length).to(device)
+    test_targets = one_hot(torch.as_tensor(np.argmax(test_data_numpy, axis=1), dtype=torch.int64), datapoint_length).to(torch.float32).to(device)
     
     # Convert to string format
     if use_word_format:
